boot v0.1

+++ RTL00_SDKV35a/component/os/freertos/freertos_v8.1.2/Source/portable/MemMang/heap_5.c
@@ -538,7 +538,7 @@ const HeapRegion_t *pxHeapRegion;
 
 	pxHeapRegion = &( pxHeapRegions[ xDefinedRegions ] );
 
-	int chip_id = HalGetChipId();
+	uint8 chip_id = HalGetChipId();
 	while( pxHeapRegion->xSizeInBytes > 0 )
 	{
 		if(pxHeapRegion->pucStartAddress > 0x20000000
